// src/shared/enums/access_permission.enum.ts
export enum AccessPermission {
  CREATE_USER = 'create_user',
  READ_USER = 'read_user',
  UPDATE_USER = 'update_user',
  DELETE_USER = 'delete_user',
  MENU_USER = 'menu_user',
  CREATE_CUSTOMER = 'create_customer',
  READ_CUSTOMER = 'read_customer',
  UPDATE_CUSTOMER = 'update_customer',
  DELETE_CUSTOMER = 'delete_customer',
  MENU_CUSTOMER = 'menu_customer',
  CREATE_UNIT_QUANTITY = 'create_unit_quantity',
  READ_UNIT_QUANTITY = 'read_unit_quantity',
  UPDATE_UNIT_QUANTITY = 'update_unit_quantity',
  DELETE_UNIT_QUANTITY = 'delete_unit_quantity',
  MENU_UNIT_QUANTITY = 'menu_unit_quantity',
  CREATE_PRODUCT = 'create_product',
  READ_PRODUCT = 'read_product',
  UPDATE_PRODUCT = 'update_product',
  DELETE_PRODUCT = 'delete_product',
  MENU_PRODUCT = 'menu_product',
}